Refactor isCourseConflict to store
