+++ src/global-firestore-data/utils.ts
@@ -16,6 +16,3 @@ export const incrementUniqueID = (amount = 1): number => {
     .set({ uniqueIncrementer: updatedID });
   return updatedID;
 };
-
-export const isCourseConflict = (uniqueId: string | number): boolean =>
-  store.state.doubleCountedCourseUniqueIDSet.has(uniqueId);
